+++ src/shared/components/location/constant/location.ts
