+++ src/shared/components/ui/button-tab/ButtonTabGroup.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import ButtonTab from '@components/button-tab/ButtonTab';
+import ButtonTab from '@ui/button-tab/ButtonTab';
 
 interface Tab {
   id: string;
@@ -29,7 +29,7 @@ export default function ButtonTabGroup({
   return (
     <div
       className={
-        'top-[4.8rem] flex flex-row gap-[1rem] border-b border-grayscale-100 bg-white px-[2rem] fixed-center'
+        'border-grayscale-100 fixed-center top-[4.8rem] flex flex-row gap-[1rem] border-b bg-white px-[2rem]'
       }
     >
       {tabs.map(tab => (
