+++ src/shared/store/regions-store.ts
@@ -1,6 +1,4 @@
 import { atom } from 'jotai';
-import type { RegionResponse } from '@/../apis/data-contracts';
+import type { RegionResponse } from 'apis/data-contracts';
 
-export const confirmedRegionsAtom = atom<Map<number, RegionResponse>>(
-  new Map()
-);
+export const regionsAtom = atom<Map<number, RegionResponse>>(new Map());
